Make POST and GET request functions more general
