+++ twitch_requests.py
@@ -3,8 +3,29 @@ from Broadcast import Broadcast
 import config
 
 
+def get_Twitch_access_token() -> str:
+    url = config.auth_Twitch_URL
+    data = {"client_id": config.client_id,
+            "client_secret": config.secret_id,
+            "grant_type": config.twitch_grant_type
+            }
+    response = dev_requests.post_request(url=url, data=data)
+    try:
+        access_token = response["access_token"]
+    except Exception:
+        access_token = ""
+
+    return access_token
+
+
 def is_broadcast_live() -> Broadcast:
-    response = dev_requests.get_request()
+    url = config.twitch_search_channels_url + config.twitch_channel_name
+    access_token = get_Twitch_access_token()
+    headers = {
+        "Client-ID": config.client_id,
+        "Authorization": "Bearer %s" % access_token
+    }
+    response = dev_requests.get_request(url=url, data=headers)
     channel_list = response["data"]
     streamer = dict()
     for channel in channel_list:
